Add --head option to match curl behaviour
The normal operation of `curl` with `-X HEAD` is
not to discard bodies, which is what a user might expect when using the
method. Instead, `curl` warns about the usage and provides the `--head`
cli flag to switch to a no-body method, and discards the body.

More information on this functionality:
https://curl.se/libcurl/c/CURLOPT_NOBODY.html