Changing naming from Manual Page to just Manual
Where it made sense in the titles, and references in the documentation,
I have updated to just `Manual` over `Manual Page` for clarity on the
purpose of the section or documentation. I have left the file naming
alone, since that was mostly used in the links and felt like the titles
might be where more confusion could be at.