Remove xmltree/indexmap dependency.
Due to xmltree re-exposing an older version of indexmap, we couldnt' upgrade to the latest version of indemap.
xmltree is a tree in-memory representation of an XML document that we use for JUnit export. As xmltree is a thin layer above xml-rs, we re-implement a thin tree in-memory XML document using xml-rs directly and remove xmltree/indexmap dependency.