Add better errors for unexpect JSON bodies
The previous code would show the entire value of the expected and actual
response bodies. For smaller bodies this is fine, but once you start
testing larger response the error becomes borderline useless.

This commit changes the error printing to show a diff between the actual
and expected bodies, if they are JSON, ignoring white space differences.
This gives the user a much better idea what the actual difference is
between the actual and expected values.

As stated this only does it for JSON, but a similar approach can be
applied to XML or any text based format, or really anything for which a
reasonable diff can be created.