Splits run function.
We begin to refcato the run function for parrallel execution. In this case, we want to reuse the bulk of the run function, while reusing the same logger instance. When run in parallel, we want to buffer log to sderr.
To do this, when we execute a file, we need to have a single (mut) instance of a logger.